BUG matrix repr printed transpose shape

+++ oomatrix/test/common.py
@@ -2,8 +2,8 @@ import numpy as np
 from nose.tools import ok_, eq_, assert_raises
 from textwrap import dedent
 
-def ndrange(shape):
-    return np.arange(np.prod(shape)).reshape(shape).astype(np.double)
+def ndrange(shape, dtype=np.double):
+    return np.arange(np.prod(shape)).reshape(shape).astype(dtype)
 
 
 def plot_add_graph(graph, max_node_size=4, block=False):
